Move BucketAsyncWrapper to abstract/bucket
