+++ pyrate_limiter/abstracts/bucket.py
@@ -113,3 +113,63 @@ def get_bucket_availability(bucket: Union[AbstractBucket], item: RateItem) -> Un
 
     assert isinstance(bound_item, RateItem)
     return _calc_availability(bound_item)
+
+
+class BucketAsyncWrapper(AbstractBucket):
+    """BucketAsyncWrapper is a wrapping over any bucket
+    that turns a async/synchronous bucket into an async one
+    """
+
+    def __init__(self, bucket: Union[AbstractBucket]):
+        assert isinstance(bucket, AbstractBucket)
+        self.bucket = bucket
+
+    async def put(self, item: RateItem):
+        result = self.bucket.put(item)
+
+        while isawaitable(result):
+            result = await result
+
+        return result
+
+    async def count(self):
+        result = self.bucket.count()
+
+        while isawaitable(result):
+            result = await result
+
+        return result
+
+    async def leak(self, current_timestamp: Optional[int] = None) -> int:
+        result = self.bucket.leak(current_timestamp)
+
+        while isawaitable(result):
+            result = await result
+
+        assert isinstance(result, int)
+        return result
+
+    async def flush(self) -> None:
+        result = self.bucket.flush()
+
+        while isawaitable(result):
+            result = await result
+
+        return None
+
+    async def peek(self, index: int) -> Optional[RateItem]:
+        item = self.bucket.peek(index)
+
+        while isawaitable(item):
+            item = await item
+
+        assert item is None or isinstance(item, RateItem)
+        return item
+
+    @property
+    def failing_rate(self):
+        return self.bucket.failing_rate
+
+    @property
+    def rates(self):
+        return self.bucket.rates
